requirements.txt
fastapi
uvicorn
PyPDF2
python-dotenv
langchain-google-genai